requirements.txt
openai
openai-whisper
faster_whisper
pygame
torch
requests
pillow
pytube
pydub
yt_dlp
youtube_transcript_api
python-dotenv
elevenlabs
anthropic
diffusers
transformers
transformers[sentencepiece]
langchain
langchain-openai
langchain-anthropic
langchain-ollama
langserve[all]
opencv-python
pyttsx3
pytesseract
pyperclip
gradio
streamlit
flask
matplotlib
numpy